Updated index name and removed composite id class
